Move register/deregister event listeners into their own module
